changing to continuous

+++ src/test/java/ca/jacob/cs6735/test/MatrixTest.java
@@ -41,10 +41,10 @@ public class MatrixTest {
     @Test
     public void testCalculateOccurrences() {
         Vector v = new Vector(new double[]{1., 2., 3., 1., 1.});
-        Map<Double, Integer> expected = new HashMap<Double, Integer>();
-        expected.put(1., 3);
-        expected.put(2., 1);
-        expected.put(3., 1);
+        Map<Integer, Integer> expected = new HashMap<Integer, Integer>();
+        expected.put(1, 3);
+        expected.put(2, 1);
+        expected.put(3, 1);
         assertEquals(expected, calculateOccurrences(v));
     }
